requirements.txt
absl-py==0.9.0
astunparse==1.6.3
attrs==19.3.0
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
fire==0.3.1
gast==0.3.3
google-auth==1.16.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
more-itertools==8.4.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.4
pluggy==0.13.1
protobuf==3.12.2
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.3
python-dateutil==2.8.1
python-speech-features==0.6
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-cpu==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.46.1
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
